--- signify-ts-test/src/workflows/eba-api-verifier-test-workflow.yaml
@@ -1,12 +1,0 @@
-workflow:
-  steps:
-    eba_api_test_ecr1:
-      id: "api_test_ecr1"
-      type: "eba_api_test"
-      aids: ["ecr-aid-1"]
-      description: "Running API test for ecr-aid-1 user"
-    # vlei_verification_test_1:
-    #   id: "vlei_verification_1"
-    #   type: "vlei_verification_test"
-    #   aids: ["ecr-aid-1"]
-    #   description: "Running Vlei Verification test for ecr-aid-1 user"
