+++ signify-ts-test/test/run-workflow.test.ts
@@ -1,17 +1,8 @@
-import { VleiIssuance } from "../src/vlei-issuance";
 import path from "path";
-import { getOrCreateClients } from "./utils/test-util";
 import { resolveEnvironment, TestEnvironment } from "./utils/resolve-env";
-import { buildAidData } from "../src/utils/handle-json-config";
-import { generate_reports } from "./report.test";
 import {
-  ApiUser,
-  getApiTestData,
   getConfig,
-  getReportGenTestData,
 } from "./utils/test-data";
-import { run_api_revocation_test, run_api_test } from "./reg-pilot-api.test";
-import { run_vlei_verification_test } from "./vlei-verification.test";
 import { runWorkflow } from "./utils/run-workflow";
 
 const fs = require("fs");
